:zap: Create lib, add benches, add CI
