+++ .github/workflows/ci.yml
@@ -1,0 +1,30 @@
+name: Cargo Build & Test
+
+on:
+  push:
+    branches:
+        -main
+  pull_request:
+    branches:
+        -main
+
+env: 
+  CARGO_TERM_COLOR: always
+
+jobs:
+  build_and_test:
+    name: Rust project - latest
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        toolchain:
+          - stable
+          - beta
+          - nightly
+    steps:
+      - uses: actions/checkout@v3
+      - run: rustup update ${{ matrix.toolchain }} && rustup default ${{ matrix.toolchain }}
+      - run: cargo fmt --check
+      - run: cargo build
+      - run: cargo test
+      - run: cargo bench # results are not really representative in ci
