+++ Cargo.toml
@@ -12,3 +12,10 @@ clap = {version = "4.3.19", features = ["cargo"]}
 geozero = {version = "0.10.0", features = ["with-wkb", "with-wkt", "geojson", "with-geo"] }
 geo = {version = "0.26.0"}
 hex = "0.4.3"
+
+[dev-dependencies]
+criterion = "0.3"
+
+[[bench]]
+name = "handle_benchmark"
+harness = false
